requirements.txt
Flask==2.3.3
SQLAlchemy==1.4.53
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
PyMySQL==1.1.0
cryptography==41.0.7
Flask-Limiter==3.5.0
redis==5.0.1
requests==2.31.0
beautifulsoup4==4.12.2